+++ Products/Poi/config.py
@@ -18,3 +18,11 @@ PSC_TRACKER_ID = 'issues'
 # issue/response text.
 ISSUE_MIME_TYPES = ('text/x-web-intelligent', 'text/plain')
 DEFAULT_ISSUE_MIME_TYPE = 'text/x-web-intelligent'
+# Patterns used for recognizing links to issues and revisions:
+ISSUE_RECOGNITION_PATTERNS = \
+    [r'\B#[1-9][0-9]*\b', r'\bissue:[1-9][0-9]*\b',
+     r'\bticket:[1-9][0-9]*\b', r'\bbug:[1-9][0-9]*\b']
+REVISION_RECOGNITION_PATTERNS = \
+    [r'\br[0-9]+\b', r'\bchangeset:[0-9]+\b', r'\B\[[0-9]+\]\B']
+# Template to use when recognizing a link to another issue:
+ISSUE_LINK_TEMPLATE = '<a href="../%(bug)s">%(linktext)s</a>'